pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "humanoid_rl"
version = "0.1.0"
description = "Multi-algorithm reinforcement learning for humanoid locomotion on MuJoCo Playground"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"

# core: just enough to train JAX algorithms on playground
dependencies = [
    "jax[cuda12]>=0.4.30",
    "flax>=0.8",
    "optax>=0.2",
    "mujoco>=3.2",
    "mujoco-mjx>=3.2",
    "playground",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
# PyTorch implementations (ppo/torch/, td3/torch/, etc.)
torch = [
    "torch>=2.2",
]
# library baselines for benchmarking
baselines = [
    "brax>=0.10",
    "stable-baselines3>=2.3",
    "sb3-contrib>=2.3",
    "gymnasium>=0.29",
]
# logging + visualization
logging = [
    "wandb>=0.16",
    "tensorboard>=2.16",
    "matplotlib>=3.8",
]
# development
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]
# everything
all = [
    "humanoid_rl[torch,baselines,logging,dev]",
]

[tool.setuptools.packages.find]
include = ["src*", "scripts*", "configs*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
